+++ assets/js/popup.js
@@ -1,8 +1,3 @@
-document.getElementById('add-model').onclick = () => {
-    document.getElementById('div-add-model').style.display = 'block';
-    document.getElementById('popup-background').style.display = 'block';
-};
-
 document.getElementById('popup-background').onclick = () => {
     document.getElementById('div-del-model').style.display = 'none';
     document.getElementById('div-add-model').style.display = 'none';
@@ -30,5 +25,54 @@ document.getElementById('option-openai').onclick = () => {
     document.getElementById('option-ollama').classList.remove('checked');
     document.getElementById('option-openai').classList.add('checked');
 };
-
 document.getElementById('option-openai').click();
+
+document.getElementById('form-add-model').addEventListener('submit', function(event) {
+    event.preventDefault();
+});
+
+document.getElementById('btn-del-cancel').onclick = () => {
+    document.getElementById('div-del-model').style.display = 'none';
+    document.getElementById('popup-background').style.display = 'none';
+};
+
+document.getElementById('btn-add-cancel').onclick = () => {
+    const form = document.getElementById('form-add-model');
+    form.reset();
+    document.getElementById('div-add-model').style.display = 'none';
+    document.getElementById('popup-background').style.display = 'none';
+};
+
+document.getElementById('btn-del-submit').onclick = () => {
+    vscode.postMessage({
+        command: 'model.delete',
+        modelData: g_toDeleteModel
+    });
+    document.getElementById('div-del-model').style.display = 'none';
+    document.getElementById('popup-background').style.display = 'none';
+};
+
+document.getElementById('btn-add-submit').onclick = () => {
+    const form = document.getElementById('form-add-model');
+    if (!form.checkValidity()) { return; }
+    const optionOllama = document.getElementById('option-ollama');
+    let modelData = {};
+    modelData['model'] = document.getElementById('i-model').value;
+    modelData['title'] = document.getElementById('i-title').value;
+    if(modelData['title'].trim() === '') { delete modelData['title']; }
+    if(optionOllama.classList.contains('checked')){
+        modelData['type'] = 'ollama';
+    }
+    else{
+        modelData['type'] = 'openai';
+        modelData['base_url'] = document.getElementById('i-base_url').value;
+        modelData['api_key'] = document.getElementById('i-api_key').value;
+    }
+    vscode.postMessage({
+        command: 'model.add',
+        modelData: JSON.stringify(modelData)
+    });
+    form.reset();
+    document.getElementById('div-add-model').style.display = 'none';
+    document.getElementById('popup-background').style.display = 'none';
+};