+++ assets/js/prototypeUtils.js
@@ -1,1 +1,0 @@
-// document.getElementById('div-add-model').style.display = 'block';